HUGE PUSH

+++ 322Mobile/322Mobile/Models/Phone.cs
@@ -5,7 +5,7 @@ namespace _322Mobile.Models
   {
     public int Id { get; set; }
     public string Name { get; set; }
-    public int Score { get; set; }
+    public decimal Score { get; set; }
     public string ImageUrl { get; set; }
     public decimal Price { get; set; }
     public string[] Providers { get; set; }
